refactor: seperate mqtt settings
