+++ apps/fastapi/app/db/postgres.py
@@ -6,7 +6,7 @@ from app.config.db import db_config
 class PostgeSQL:
     def __init__(self):
         self.engine = create_engine(
-            db_config.POSTGRES_URL,
+            db_config.DATABASE_URL,
             pool_pre_ping=True,
             pool_size=10,
             max_overflow=20,
